+++ src/routes/search/Search.module.css
@@ -9,41 +9,46 @@
 
 .content {
   text-align: center;
-  /* margin-top: 3%;
-  margin-right: 9.6%;
-  margin-left: 9.6%;
-  margin-bottom: 9%; */
-  width: 1060px;
+  height: auto;
+  width: 73%;
+  max-width: 1060px;
 }
 
-.categories {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, 30%);
-  grid-gap: 1rem;
-  justify-content: center;
-}
 
-.headerMargin {
-  padding-bottom: 40px;
+.searchContainer {
+  position: relative;
+  width: 100%;
+  height: 46px;
+  margin-top: 45px;
 }
 
 .searchBar {
   width: 100%;
-  height: 35px;
+  height: 100%;
   border: 0px solid #000000;
   background-color: #f3f3f3;
   border-radius: 8px;
+  font-size: 18px;
 
-  padding: 10px 40px;
+  padding: 0 48px;
 }
 
 .searchBar::placeholder {
   color: #686868;
 }
 
+.searchIcon {
+  content: "";
+  position: absolute;
+  top: 12px;
+  left: 16px;
+  width: 20px;
+}
+
 .bigItemList {
   padding: 4px 0;
-  grid-gap: 5%;
+  grid-column-gap: 5%;
+  grid-row-gap: 30px;
   display: flex;
   flex-wrap: wrap;
   justify-content: space-between;
@@ -51,7 +56,6 @@
 }
 
 .bigItemList li {
-  margin-top: 50px;
   display: inline-block;
   width: 19.4%;
   max-width: 200px;
@@ -59,22 +63,15 @@
   height: auto;
 }
 
-.searchIcon {
-  content: "";
-  position: relative;
-  bottom: 0;
-  width: 20px;
-  float: left;
-  top: 30px;
-  left: 10px;
-}
 
 .resultsText {
+  padding: 30px 0;
   text-align: left;
   font-size: 18px;
-  line-height: 27px;
   font-weight: 700;
 }
+
+
 
 .noResultsContainer p {
   font-size: 16px;
@@ -85,7 +82,6 @@
   font-size: 18px;
   line-height: 27px;
   font-weight: 700;
-  margin-top: 20px;
 }
 
 .pickupTextField {
@@ -108,40 +104,72 @@
 }
 
 .categories {
-  margin-top: 50px;
+  width: 100%;
+  height: auto;
+  margin-top: 45px;
+  padding-bottom: 140px;
   display: flex;
   flex-wrap: wrap;
+  grid-column-gap: 18px;
   /* display: grid;
   grid-template-columns: repeat(auto-fill, 30%); */
-  grid-column-gap: 0.4rem;
+  /* grid-column-gap: 0.4rem; */
   grid-row-gap: 3rem;
   justify-content: space-between;
 }
 
-@media only screen and (max-width: 768px) {
-  .container {
-    margin-right: 3%;
-    margin-left: 3%;
+@media only screen and (max-width: 900px) {
+
+  .content {
+    width: 91%;
   }
 
+  .searchContainer {
+    position: relative;
+    width: 100%;
+    height: 35px;
+    margin-top: 60px;
+  }
+  
   .searchBar {
-    width: 90%;
-    margin-right: 3%;
+    width: 100%;
+    height: 46px;
+    border: 0px solid #000000;
+    background-color: #f3f3f3;
+    border-radius: 8px;
+    font-size: 18px;
+  
+    padding: 12px 48px;
   }
-
+  
+  .searchBar::placeholder {
+    color: #686868;
+  }
+  
   .searchIcon {
+    content: "";
     position: absolute;
-    top: 57px;
-    left: 55px;
+    top: 12px;
+    left: 16px;
+    width: 20px;
   }
 
   .categories {
-    grid-template-columns: repeat(auto-fill, 40%); /* 2 */
+    margin-top: 25px;
+    grid-row-gap: 20px;
+  }
+}
+
+@media only screen and (max-width: 700px) {
+
+  .bigItemList {
+    justify-content: space-evenly;
   }
 
-  .headerMargin {
-    margin-left: 5%;
-    margin-right: 2%;
-    padding-top: 40px;
+  .bigItemList li {
+    width: 40%;
+    max-width: 320px;
+    min-width: 170px;
   }
+
 }
