+++ src/shared/category-menu/CategoryMenu.js
